pyproject.toml
[project]
name = "newsscraper"
version = "0.1.0"
description = "API for scraping and analyzing news articles with NLP keyword filtering"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "News Scraper Team"}
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "lxml_html_clean>=0.1.0",
    "selenium>=4.15.2",
    "newspaper3k>=0.2.8",
    "patchright>=1.0.0",
    "spacy>=3.7.2",
    "nltk>=3.8.1",
    "python-dateutil>=2.8.2",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.1",
    "async-timeout>=4.0.3",
    "streamlit>=1.29.0",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
]
